feat : 선택된 코인 parameter로 전달 구현

+++ BithumbProject/BithumbProject/Sources/SupportingFiles/Extensions/String+Extension.swift
@@ -30,10 +30,10 @@ extension String {
     }
     
     var decimal: String? {
-        guard let number = Int(self) else { return nil }
+        guard let number = Double(self) else { return nil }
         let formatter = NumberFormatter()
         formatter.numberStyle = .decimal
-        return formatter.string(for: Int(number))
+        return formatter.string(for: Double(number))
     }
     
     var rounded: String? {
